% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-corpora.R
\docType{data}
\name{data_corpus_debates}
\alias{data_corpus_debates}
\title{2020 US Presidential Debates}
\format{
The docname is the date of the debate.  The corpus docvars consist of:
\describe{
\item{location}{factor; location of the debate (Cleveland and Nashville).}
\item{date}{date; date of the debate (2020-09-29 and 2020-10-22).}
}
}
\source{
\url{https://www.presidency.ucsb.edu}
}
\usage{
data_corpus_debates
}
\description{
A corpus object containing the transcripts of the two US Presidential Debates
Donald J. Trump and Joe Biden. Each document includes the full debate text by the two
politicians and the moderator.
}
\examples{
# transform debate-level corpus to level of statements
library(quanteda)
data_corpus_debatesseg <- corpus_segment(data_corpus_debates,
                                         pattern =  "\\\\s*[[:upper:]]+:\\\\s+",
                                         valuetype = "regex",
                                         case_insensitive = FALSE)
}
\references{
"The American Presidency Project"
}
\keyword{data}
