% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-corpora.R
\docType{data}
\name{data_corpus_TAhotels}
\alias{data_corpus_TAhotels}
\title{TripAdvisor Hotel Reviews}
\format{
The only corpus docvar consists of \code{Rating}, ranging from 1 to 5.
}
\source{
\url{https://www.kaggle.com/andrewmvd/trip-advisor-hotel-reviews}
}
\usage{
data_corpus_TAhotels
}
\description{
A corpus object containing 20k TripAdvisor hotel reviews, and ratings.
}
\section{License}{
 \href{https://creativecommons.org/licenses/by-nc/4.0/}{Attribution-NonCommercial 4.0 International (CC BY-NC 4.0)}
}

\references{
Alam, M. H., Ryu, W.-J., Lee, S. (2016). Joint multi-grain topic
sentiment: modeling semantic aspects for online reviews. \emph{Information
Sciences} 339, 206–223. \doi{10.1016/j.ins.2016.01.013}
}
\keyword{data}
