% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-corpora.R
\docType{data}
\name{data_corpus_trumptweets}
\alias{data_corpus_trumptweets}
\title{Trump Twitter archive}
\format{
The docname is the ID of the original Tweet.  The corpus docvars consist of:
\describe{
\item{isRetweet}{logical; if \emph{TRUE}, the Tweet was a retweet.}
\item{isDeleted}{logical; if \emph{TRUE}, the Tweet was deleted.}
\item{device}{factor; the device from which the Tweet was sent.}
\item{favorites}{integer; the total number of favorites the Tweet received.}
\item{retweets}{integer; total number of times the Tweet was retweeted.}
\item{isFlagged}{logical; if \emph{TRUE}, the Tweet was flagged.}
\item{datetime}{datetime format in UTC.}
}
}
\source{
\url{https://www.thetrumparchive.com}
}
\usage{
data_corpus_trumptweets
}
\description{
A corpus object containing a dataset consisting of all of Trump's Tweets,
going back to 2009.  Total: 56,571.
}
\keyword{data}
